# openapi.yaml
openapi: 3.0.3
info:
  title: storeit — backend
  contact:
    email: evgeny@evseevs.ru
  version: 0.0.1
servers:
  - url: http://localhost:8080/api/v0
  # - url: https://{orgName}.storeit.ru/api/v0
  #   variables:
  #     orgName:
  #       default: demo
  #       description: Organization name
...
paths:
  /orgs:
    get:
      tags:
        - org
      summary: Get list of Organizations
      operationId: getOrgs
      parameters:
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PagedResponse"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/OrgWithAudit"
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - org
      summary: Create Organization
      operationId: createOrg
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Org"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Org"
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /orgs/{id}:
    parameters:
      - name: id
        in: path
        description: Organization ID
        required: true
        schema:
          type: string
          format: uuid
    put:
      tags:
        - org
      summary: Update Organization
      operationId: updateOrg
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Org"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Org"
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - org
      summary: Delete Organization
      operationId: deleteOrg
      responses:
        "200":
          description: Successful operation
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags:
        - org
      summary: Get Organization by ID
      operationId: getOrgById
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Org"
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /units:
    get:
      tags:
        - unit
      summary: Get list of Organization Units
      operationId: getUnits
      parameters:
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PagedResponse"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/Unit"
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - unit
      summary: Create Organization Unit
      operationId: createUnit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Unit"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
        default:
          description: General Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
    OffsetParam:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        minimum: 0
        default: 0
        format: int32
    LimitParam:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
        format: int32
  schemas:
    AuditFields:
      type: object
      properties:
        created_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          example: "2024-01-20T18:00:00Z"
        created_by:
          type: string
          format: uuid
          readOnly: true
          nullable: true
          description: ID of the employee who created the entity
          example: "def3df1a-7b8f-4552-b437-a1eab851403f"
        updated_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
          example: "2024-01-21T10:30:00Z"
        updated_by:
          type: string
          format: uuid
          readOnly: true
          nullable: true
          description: ID of the employee who last updated the entity
          example: "123e4567-e89b-12d3-a456-426614174001"
      required:
        - created_at
        - created_by
        - updated_at
        - updated_by
    OrgWithAudit:
      type: object
      allOf:
        - $ref: "#/components/schemas/AuditFields"
        - $ref: "#/components/schemas/Org"
    Org:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          example: def3df1a-7b8f-4552-b437-a1eab851403f
        name:
          type: string
          example: Exotic
        subdomain:
          type: string
          example: exotic
      required:
        - name
        - subdomain
    Error:
      description: Represents error object
      type: object
      properties:
        code:
          type: integer
          format: int64
        message:
          type: string
        details:
          type: object
          nullable: true
          example: { "id": "123e4567-e89b-12d3-a456-426614174000" }
      required:
        - code
        - message
        - details
    Unit:
      type: object
      required:
        - name
      allOf:
        - $ref: "#/components/schemas/AuditFields"
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          example: def3df1a-7b8f-4552-b437-a1eab851403f
        name:
          type: string
          minLength: 1
          maxLength: 100
          pattern: "^[\\w\\s-]+$"
          example: Moscow
        address:
          type: string
          nullable: true
          maxLength: 200
          example: Amurskaya St. 3
...
    PaginationMetadata:
      type: object
      properties:
        offset:
          type: integer
          example: 0
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 50
    PagedResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        metadata:
          $ref: "#/components/schemas/PaginationMetadata"